+++ src/modules/remote/index.ts
@@ -1,0 +1,1 @@
+export { RemotePanel } from "./RemotePanel";